pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "wakalat-ai-backend"
version = "1.0.0"
description = "MCP Server for Wakalat-AI Legal Assistant - Provides tools and resources for legal assistance to lawyers in India"
authors = [
    { name = "Wakalat-AI Team", email = "team@wakalat-ai.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Legal Industry",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # MCP Server Dependencies
    "mcp>=1.0.0",
    # HTTP Server Dependencies
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    # Vector Database for semantic search
    "chromadb>=0.4.0",
    # LLM Integration
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    # Gemini + Firecrawl for Deep Research
    "google-genai>=1.0.0",
    "firecrawl-py>=1.0.0",
    # Document Processing
    "pypdf2>=3.0.0",
    "python-docx>=1.0.0",
    "pdfplumber>=0.10.0",
    # Text Processing
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    # Embeddings
    "sentence-transformers>=2.2.0",
    # PostgreSQL for vector search
    "psycopg2-binary>=2.9.0",
    # Web Scraping (for fetching case laws)
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    # Indian Legal System specific
    "indic-nlp-library>=0.92",
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    # Logging and Monitoring
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    
    # Code Quality
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
web = [
    # Optional web interface
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
wakalat-ai = "src.server:main"
wakalat-ai-stdio = "src.server:main"
wakalat-ai-http = "src.http_server:main"

[project.urls]
Homepage = "https://github.com/BE-project-vesit/Wakalat-AI-Backend"
Repository = "https://github.com/BE-project-vesit/Wakalat-AI-Backend"
Issues = "https://github.com/BE-project-vesit/Wakalat-AI-Backend/issues"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_unimported = true
no_implicit_optional = true
warn_return_any = true
warn_unused_configs = true
exclude = [
    "venv",
    ".venv",
    "env",
    ".env",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
